PeriodicValue: Update JSON format
The JSON message for periodic values would contain the periodic sensor-value
and epoch timestamp (in milliseconds). The key for periodic value of any
sensor would be its tag name according to the IO list document (PR-190419-IOL).
The key for epoch timestamp would be the same as that of the value preceeded
by “t_”. For example:

{"PT_61":61.4820,"t_PT_61":1587638098940}

JIRA: https://intechww.atlassian.net/browse/PAR-131

Signed-off-by: Fahad Usman <fahad.usman@intechww.com>