MqttHandler: Explicitly clear connected flag if connection is lost
Sometimes MQTT connection is broken but connection_lost callback is not
invoked, in that case we clear the connected_ flag of the callback class
so that subsequent reconnection attempt can be made. This is a hack and
the proper solution would be to ensure that connection_lost callback is
always called when connection to the broker is lost.

JIRA: https://intechww.atlassian.net/browse/PAR-154

Signed-off-by: Fahad Usman <fahad.usman@intechww.com>